Extract loadLessons and supporting fns into its own module
